requirements.txt
numpy
pillow
matplotlib
tqdm
chromadb
open-clip-torch
